Split up class into multiple files
