+++ src/column.ts
@@ -1,0 +1,9 @@
+export class Column {
+    id: number;
+    mode: string;
+
+    constructor(id: number, mode: string) {
+        this.id = id;
+        this.mode = mode;
+    }
+}
